requirements.txt
fastapi
uvicorn
tensorflow==2.16.1
numpy
Pillow
python-multipart
opencv-python-headless